Confirmation page
Added a 2nd scratch file (confirmation-DEB2.html) and played around
with adding icons in each of the tabs, scrapping the email tab and
creating a separate one for each type of social media.

It also appears that if you’re posting something to either FB, Twitter
or Pintrest the form is already branded. Unless there’s a way around
that…?